requirements.txt
click==8.0.3
gunicorn==20.1.0
beautifulsoup4==4.11.1
bs4==0.0.1
Flask==2.1.2
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.4.0
mal-api
MarkupSafe==2.1.1
numpy==1.23.0
pandas==1.4.3
Pillow==9.1.0
requests==2.28.1
scikit-learn==1.1.1
scipy==1.8.1
sklearn==0.0
Werkzeug==2.1.2
bz2file==0.98
